+++ Kinvey-Xamarin/Realtime/Stream.cs
@@ -193,7 +193,7 @@ namespace Kinvey
 			var urlParameters = new Dictionary<string, string>();
 			urlParameters.Add(Constants.STR_APP_KEY, ((KinveyClientRequestInitializer)KinveyClient.RequestInitializer).AppKey);
 			urlParameters.Add(Constants.STR_REALTIME_STREAM_NAME, StreamName);
-			urlParameters.Add("userID", KinveyClient.ActiveUser.Id);
+			urlParameters.Add("userID", receiverID);
 
 			var requestStreamPublish = new StreamPublishAccessRequest(receiverID, KinveyClient, urlParameters);
 			KinveyClient.InitializeRequest(requestStreamPublish);
