+++ TestApps/DemoKLS/DemoKLS/ViewController.cs
@@ -54,6 +54,7 @@ namespace DemoKLS
 			};
 			View.AddSubview(passwordField);
 
+
 			var buttonWidth = (w / 2) - 20;
 			buttonLogin = UIButton.FromType(UIButtonType.System);
 			buttonLogin.Frame = new CGRect(10, 162, w - 20, 44);
@@ -62,22 +63,12 @@ namespace DemoKLS
 			buttonLogin.BackgroundColor = colorBackgroundButtonLogin;
 
 			buttonLogin.TouchUpInside += async (sender, e) => {
-				AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-				await myAppDel.Login(usernameField.Text, passwordField.Text);
+				//AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
+				//await myAppDel.Login(usernameField.Text, passwordField.Text);
 			};
 
 			View.AddSubview(buttonLogin);
 
-			//buttonLoginAuto = UIButton.FromType(UIButtonType.System);
-			//buttonLoginAuto.Frame = new CGRect(w - buttonWidth - 10, 200, buttonWidth, 44);
-			//buttonLoginAuto.SetTitle("Login Test", UIControlState.Normal);
-			//buttonLoginAuto.SetTitleColor(UIColor.White, UIControlState.Normal);
-			//buttonLoginAuto.BackgroundColor = colorBackgroundButtonLogin;
-
-			//buttonLoginAuto.TouchUpInside += async (sender, e) => {
-			//	AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-			//	await myAppDel.Login("Test", "test");
-			//};
 
 			buttonLoginAlice = UIButton.FromType(UIButtonType.System);
 			buttonLoginAlice.Frame = new CGRect(10, 250, w - 20, 44);
@@ -118,259 +109,24 @@ namespace DemoKLS
 
 			//View.AddSubview(buttonLoginCharlie);
 
-			//buttonLoginDan = UIButton.FromType(UIButtonType.System);
-			//buttonLoginDan.Frame = new CGRect(10, 400, w - 20, 44);
-			//buttonLoginDan.SetTitle("Login Dan", UIControlState.Normal);
-			//buttonLoginDan.SetTitleColor(UIColor.White, UIControlState.Normal);
-			//buttonLoginDan.BackgroundColor = colorBackgroundButtonLogin;
+			buttonLoginDan = UIButton.FromType(UIButtonType.System);
+			buttonLoginDan.Frame = new CGRect(10, 400, w - 20, 44);
+			buttonLoginDan.SetTitle("Login Dan", UIControlState.Normal);
+			buttonLoginDan.SetTitleColor(UIColor.White, UIControlState.Normal);
+			buttonLoginDan.BackgroundColor = colorBackgroundButtonLogin;
 
-			//buttonLoginDan.TouchUpInside += async (sender, e) => {
-			//	AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-			//	await myAppDel.LoginDan();
-			//};
+			buttonLoginDan.TouchUpInside += async (sender, e) => {
+				AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
+				await myAppDel.LoginDan();
+			};
 
-			//View.AddSubview(buttonLoginDan);
+			View.AddSubview(buttonLoginDan);
 		}
 
 		public override void DidReceiveMemoryWarning()
 		{
 			base.DidReceiveMemoryWarning();
 			// Release any cached data, images, etc that aren't in use.
-		}
-	}
-
-	public partial class PatientViewController : UIViewController
-	{
-		internal UITextField SenderIDView;
-		internal UITextField MessageView;
-		nfloat h = 31.0f;
-		UIColor colorBackgroundButtonLogin = UIColor.FromRGB(5, 58, 114);
-		//UIColor colorBackgroundButtonLogin = UIColor.FromRGB(92, 127, 159);
-		UIColor colorDarkBlue = UIColor.FromRGB(7, 69, 126);
-		UIColor colorLightBlue = UIColor.FromRGB(92, 127, 159);
-
-		public override void ViewDidLoad()
-		{
-			base.ViewDidLoad();
-
-			Title = "Demo Kinvey Live Service - Patient";
-			View.BackgroundColor = UIColor.FromRGB(7, 69, 126);
-			nfloat w = View.Bounds.Width;
-
-			AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-
-			//SenderIDView = new UITextField
-			//{
-			//	Placeholder = "Sender ID",
-			//	Frame = new CGRect(10, 82, w - 20, h),
-			//	BorderStyle = UITextBorderStyle.RoundedRect,
-			//	BackgroundColor = UIColor.White,
-			//	TextColor = UIColor.Black
-			//};
-
-			//View.AddSubview(SenderIDView);
-
-			MessageView = new UITextField
-			{
-				Placeholder = "Message",
-				Frame = new CGRect(10, 82, w - 20, h),
-				BorderStyle = UITextBorderStyle.RoundedRect,
-				BackgroundColor = UIColor.White,
-				TextColor = UIColor.Black
-			};
-
-			View.AddSubview(MessageView);
-
-			//UITextField PublishMessageView = new UITextField
-			//{
-			//	Placeholder = "Message to Publish",
-			//	BorderStyle = UITextBorderStyle.RoundedRect,
-			//	Frame = new CGRect(10, 202, w - 20, h),
-			//};
-
-			//View.AddSubview(PublishMessageView);
-
-			//UIButton buttonPublish;
-			//buttonPublish = UIButton.FromType(UIButtonType.System);
-			//buttonPublish.Frame = new CGRect(10, 242, w - 20, 44);
-			//buttonPublish.SetTitle("Publish", UIControlState.Normal);
-			//buttonPublish.SetTitleColor(UIColor.Black, UIControlState.Normal);
-			//buttonPublish.BackgroundColor = UIColor.Gray;
-			//buttonPublish.TouchUpInside += async (sender, e) => {
-			//	await myAppDel.PublishStatus(PublishMessageView.Text);
-			//	PublishMessageView.Text = String.Empty;
-			//};
-
-			//View.AddSubview(buttonPublish);
-
-			UIButton buttonLogout;
-			buttonLogout = UIButton.FromType(UIButtonType.System);
-			buttonLogout.Frame = new CGRect(10, 322, w - 20, 44);
-			buttonLogout.SetTitle("Logout", UIControlState.Normal);
-			buttonLogout.SetTitleColor(UIColor.Black, UIControlState.Normal);
-			buttonLogout.BackgroundColor = UIColor.Gray;
-
-			var user = new UIViewController();
-			user.View.BackgroundColor = colorDarkBlue;
-
-			buttonLogout.TouchUpInside += async (sender, e) => {
-				await myAppDel.Logout();
-			};
-
-			View.AddSubview(buttonLogout);
-		}
-
-		public override void DidReceiveMemoryWarning()
-		{
-			base.DidReceiveMemoryWarning();
-		}
-
-		public void ChangeText(string sender, string msg)
-		{
-			//SenderIDView.Frame = new CGRect(10, 82, View.Bounds.Width - 20, h);
-			//SenderIDView.Text = "Sender ID: " + sender;
-
-			MessageView.Frame = new CGRect(10, 82, View.Bounds.Width - 20, h);
-			MessageView.Text = "Message: " + msg;
-		}
-	}
-
-	public partial class DoctorViewController : UIViewController
-	{
-		internal UITextField SenderIDView;
-		internal UITextField MessageView;
-		internal UITextField TimeView;
-
-		nfloat h = 31.0f;
-		//UIColor colorBackgroundButtonLogin = UIColor.FromRGB(5, 58, 114);
-		UIColor colorBackgroundButtonLogin = UIColor.FromRGB(92, 127, 159);
-		UIColor colorDarkBlue = UIColor.FromRGB(7, 69, 126);
-		UIColor colorLightBlue = UIColor.FromRGB(92, 127, 159);
-
-		public override void ViewDidLoad()
-		{
-			base.ViewDidLoad();
-
-			Title = "Demo Kinvey Live Service - Doctor";
-			View.BackgroundColor = UIColor.FromRGB(7, 69, 126);
-			nfloat w = View.Bounds.Width;
-			var buttonWidth = (w / 2) - 20;
-
-			AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-
-			//SenderIDView = new UITextField
-			//{
-			//	Placeholder = "Sender ID",
-			//	Frame = new CGRect(10, 82, w - 20, h),
-			//	BorderStyle = UITextBorderStyle.RoundedRect,
-			//	BackgroundColor = UIColor.White,
-			//	TextColor = UIColor.Black
-			//};
-
-			//View.AddSubview(SenderIDView);
-
-			MessageView = new UITextField
-			{
-				Placeholder = "Message",
-				Frame = new CGRect(10, 82, w - 20, h),
-				BorderStyle = UITextBorderStyle.RoundedRect,
-				BackgroundColor = UIColor.White,
-				TextColor = UIColor.Black
-			};
-
-			View.AddSubview(MessageView);
-
-			TimeView = new UITextField
-			{
-				Placeholder = "Roundtrip Time",
-				Frame = new CGRect(10, 122, w - 20, h),
-				BorderStyle = UITextBorderStyle.RoundedRect,
-				BackgroundColor = UIColor.White,
-				TextColor = UIColor.Black
-			};
-
-			View.AddSubview(TimeView);
-
-			//UITextField PublishMessageView = new UITextField
-			//{
-			//	Placeholder = "Message to Publish",
-			//	BorderStyle = UITextBorderStyle.RoundedRect,
-			//	Frame = new CGRect(10, 202, w - 20, h),
-			//};
-
-			//View.AddSubview(PublishMessageView);
-
-			UIButton buttonPublishDecrement;
-			buttonPublishDecrement = UIButton.FromType(UIButtonType.System);
-			buttonPublishDecrement.Frame = new CGRect(10, 202, buttonWidth, 44);
-			buttonPublishDecrement.SetTitle("Decrement", UIControlState.Normal);
-			buttonPublishDecrement.SetTitleColor(UIColor.Red, UIControlState.Normal);
-			buttonPublishDecrement.BackgroundColor = colorLightBlue;
-			//buttonPublishDecrement.BackgroundColor = UIColor.Red;
-			buttonPublishDecrement.TouchUpInside += async (sender, e) => {
-				await myAppDel.PublishCommand(MedicalDeviceCommand.EnumCommand.DECREMENT);
-				//PublishMessageView.Text = String.Empty;
-			};
-
-			View.AddSubview(buttonPublishDecrement);
-
-			UIButton buttonPublishIncrement;
-			buttonPublishIncrement = UIButton.FromType(UIButtonType.System);
-			buttonPublishIncrement.Frame = new CGRect(w - buttonWidth - 10, 202, buttonWidth, 44);
-			buttonPublishIncrement.SetTitle("Increment", UIControlState.Normal);
-			buttonPublishIncrement.SetTitleColor(UIColor.Green, UIControlState.Normal);
-			buttonPublishIncrement.BackgroundColor = colorLightBlue;
-			//buttonPublishIncrement.BackgroundColor = UIColor.Green;
-			buttonPublishIncrement.TouchUpInside += async (sender, e) => {
-				await myAppDel.PublishCommand(MedicalDeviceCommand.EnumCommand.INCREMENT);
-				//PublishMessageView.Text = String.Empty;
-			};
-
-			View.AddSubview(buttonPublishIncrement);
-
-			//buttonLoginAuto = UIButton.FromType(UIButtonType.System);
-			//buttonLoginAuto.Frame = new CGRect(w - buttonWidth - 10, 200, buttonWidth, 44);
-			//buttonLoginAuto.SetTitle("Login Test", UIControlState.Normal);
-			//buttonLoginAuto.SetTitleColor(UIColor.White, UIControlState.Normal);
-			//buttonLoginAuto.BackgroundColor = colorBackgroundButtonLogin;
-
-			//buttonLoginAuto.TouchUpInside += async (sender, e) => {
-			//	AppDelegate myAppDel = (UIApplication.SharedApplication.Delegate as DemoKLS.AppDelegate);
-			//	await myAppDel.Login("Test", "test");
-			//};
-
-			UIButton buttonLogout;
-			buttonLogout = UIButton.FromType(UIButtonType.System);
-			buttonLogout.Frame = new CGRect(10, 322, w - 20, 44);
-			buttonLogout.SetTitle("Logout", UIControlState.Normal);
-			buttonLogout.SetTitleColor(UIColor.Black, UIControlState.Normal);
-			buttonLogout.BackgroundColor = UIColor.Gray;
-
-			var user = new UIViewController();
-			user.View.BackgroundColor = UIColor.FromRGB(7, 69, 126);
-
-			buttonLogout.TouchUpInside += async (sender, e) => {
-				await myAppDel.Logout();
-			};
-
-			View.AddSubview(buttonLogout);
-		}
-
-		public override void DidReceiveMemoryWarning()
-		{
-			base.DidReceiveMemoryWarning();
-		}
-
-		public void ChangeText(string msg, string time)
-		{
-			//SenderIDView.Frame = new CGRect(10, 82, View.Bounds.Width - 20, h);
-			//SenderIDView.Text = "Sender ID: " + sender;
-
-			MessageView.Frame = new CGRect(10, 82, View.Bounds.Width - 20, h);
-			MessageView.Text = "Message: " + msg;
-
-			TimeView.Frame = new CGRect(10, 122, View.Bounds.Width - 20, h);
-			TimeView.Text = "Roundtrip Time: " + time;
 		}
 	}
 }
